Small code refactoring
Split the `registry.rs` file into smaller units. This improves
maintenance and extendability.
